+++ app/controllers/adverse_event_comments_controller.rb
@@ -1,3 +1,4 @@
+# Project editors and owners can discuss an adverse event
 class AdverseEventCommentsController < ApplicationController
   before_action :authenticate_user!