+++ app/helpers/adverse_event_files_helper.rb
@@ -1,0 +1,2 @@
+module AdverseEventFilesHelper
+end