+++ db/migrate/20150930174940_create_adverse_event_files.rb
@@ -1,0 +1,16 @@
+class CreateAdverseEventFiles < ActiveRecord::Migration
+  def change
+    create_table :adverse_event_files do |t|
+      t.integer :project_id
+      t.integer :adverse_event_id
+      t.integer :user_id
+      t.string :attachment
+
+      t.timestamps null: false
+    end
+
+    add_index :adverse_event_files, :project_id
+    add_index :adverse_event_files, :adverse_event_id
+    add_index :adverse_event_files, :user_id
+  end
+end